fix(passport): Rollup's favicon still displayed on the MetaMask Notification window post enabling custom domain and white labelling
